Extract and refactor main class
