+++ src/main/java/com/github/codeidoscope/InputValidator.java
@@ -1,0 +1,17 @@
+package com.github.codeidoscope;
+
+public class InputValidator {
+
+    String getInputAndPrepareOutput(String inputLine) {
+
+        if (inputLine.equalsIgnoreCase("MARCO")) {
+            return "POLO";
+        }
+
+        if (inputLine.equalsIgnoreCase("ECHO")) {
+            return "ECHO";
+        }
+
+        return "";
+    }
+}
